requirements.txt
textual>=0.40.0
google-generativeai
openai
httpx
python-dotenv
